+++ packages/core/index.ts
@@ -5,7 +5,6 @@ export * from './directive';
 export * from './indiv';
 export * from './lifecycle';
 export * from './nv-module';
-export * from './render';
 export * from './types';
 export * from './utils';
 export * from './vnode';