+++ packages/platform-server/plugins/to-string.ts
@@ -2,52 +2,37 @@ import { InDiv, INvModule } from '@indiv/core';
 import { TRouter, nvRouteStatus } from '@indiv/router';
 import { _document } from '../renderer';
 import { PlatformServer } from './platform-server';
-// import { buildPath, generalDistributeRoutes } from '../router';
+import { buildPath, generalDistributeRoutes } from '../router';
 
-async function getDOMString(): Promise<string> {
-  return _document.getElementById('root').innerHTML;
-}
-
-async function renderApp(rootModule: Function, url?: string, nvRootPath?: string) {
+/**
+ * render a Indiv app to string
+ * 
+ * if has routeConfig, will await route render
+ *
+ * @export
+ * @param {Function} rootModule
+ * @param {string} [url]
+ * @param {{ routes?: TRouter[], nvRootPath?: string }} [routeConfig]
+ * @returns {Promise<string>}
+ */
+export async function renderToString(rootModule: Function, url?: string, routeConfig?: { routes?: TRouter[], nvRootPath?: string }): Promise<string> {
+  if (_document.getElementById('root')) _document.getElementById('root').innerHTML = '';
   const inDiv = new InDiv();
   inDiv.bootstrapModule(rootModule);
   inDiv.use(PlatformServer);
   await inDiv.init();
-  nvRouteStatus.nvRootPath = nvRootPath;
-  nvRouteStatus.nvRouteObject = {
-    path: url,
-    query: null,
-    data: null,
-  };
-}
-
-// todo async await
-/**
- * render a Indiv app to string
- *
- * @export
- * @param {InDiv} indiv
- * @param {string} [url]
- * @param {TRouter[]} [routes]
- * @returns {Promise<string>}
- */
-export async function renderToString(rootModule: Function, url?: string, nvRootPath?: string): Promise<string> {
-  if (_document.getElementById('root')) _document.getElementById('root').innerHTML = '';
-  await renderApp(rootModule, url, nvRootPath);
-  // setTimeout(() => {
-  //   renderApp(rootModule, url, nvRootPath);
-  // }, 0);
-  // const inDiv = new InDiv();
-  // inDiv.bootstrapModule(rootModule);
-  // inDiv.use(PlatformServer);
-  // await inDiv.init();
-  // nvRouteStatus.nvRootPath = nvRootPath;
-  // nvRouteStatus.nvRouteObject = {
-  //   path: url,
-  //   query: null,
-  //   data: null,
-  // };
-  const returnString = _document.getElementById('root').innerHTML;
-  // const returnString = await getDOMString();
-  return returnString.replace(/^(\<div\>)/g, '').replace(/(\<\/div\>$)/g, '');
+  if (url && routeConfig) {
+    nvRouteStatus.nvRootPath = routeConfig.nvRootPath;
+    nvRouteStatus.nvRouteObject = {
+      path: url,
+      query: null,
+      data: null,
+    };
+    const renderRouteList = buildPath(url);
+    const routesList: TRouter[] = [];
+    const loadModuleMap: Map<string, INvModule> = new Map();
+    await generalDistributeRoutes(routeConfig.routes, routesList, renderRouteList, inDiv, loadModuleMap);
+  }
+  const content = _document.getElementById('root').innerHTML;
+  return content.replace(/^(\<div\>)/g, '').replace(/(\<\/div\>$)/g, '');
 }